refactor: 당첨 결과 배열 자료형 재정의
- 보너스 결과 배열과의 타입 일치를 위해 당첨 결과를 담는 배열의 자료형을 int로 변경합니다.

+++ src/main/java/lotto/service/ResultGenerator.java
@@ -7,7 +7,7 @@ import lotto.domain.Lotto;
 
 public class ResultGenerator {
 
-    private final List<Long> winningResult = new ArrayList<>();
+    private final List<Integer> winningResult = new ArrayList<>();
     private final List<Integer> bonusResult = new ArrayList<>();
 
     public ResultGenerator(List<Lotto> lottoTicket, Lotto winning, Bonus bonus) {
@@ -20,7 +20,7 @@ public class ResultGenerator {
     }
 
     private void countWinning(Lotto lotto, Lotto winning) {
-        winningResult.add(lotto.getNumbers().stream()
+        winningResult.add((int) lotto.getNumbers().stream()
                 .filter(number -> winning.getNumbers().contains(number))
                 .count());
     }
@@ -37,7 +37,7 @@ public class ResultGenerator {
         bonusResult.add(0);
     }
 
-    public List<Long> getWinningResult() {
+    public List<Integer> getWinningResult() {
         return winningResult;
     }
